Several things - read description (home page/graphs/graph.js files
Amended getGraphData.js and created getMyGraphData.js - the latter getting data based on the uid_source field in firebase.
getGraphData could be used on the home page or on something similar - like a dashboard.
Changed home page so that it looks nicer imo:
 - added some element plus rows and columns
 - css stylings
 - changed some syntax with the text-blocks
 - added several classes for css etc
 - changed the login and register buttons from a router-link to use functions and router.push('<register or login>') method
header.vue - changed the logout button so it follows the same router.push. with the "redirectToLogin" function
FYI - you can put as many things as you want in an v-on:click / @click event listener - see line 27
Got rid of a lot of commented out stuff in myGraphs