Big DATA changes
Changed the upload form to match what was needed in firebase for new apex charts - forces firebase to store it in a nice format for apex.
Changed mygraphs and data to acoomodate this.
databytags is a work in progress ==> each of these have a related js file that has also been changed or is a wip

Added in routes for the two new graph pages [Data, DataByTags]

added a space into header.vue because it was annoying