Data pages working (mostly) READTHIS
Added "data" (for all graphs stored), finalized "mygraphs" (only user-uploaded data using FK of uid_source) and implemented search firebase graph data page "By Tags" which you will see in the header
I had a bug where mygraphs would load before the firebase authentication would trigger, so it would fail and bring up 0 results - to my knowledge this is fixed, but could be a problem depending on browser, pc speed etc.

BY TAGS: This page is not finished. You can search, but the result set is incremental - result array is incremented instead of replaced when a second search is performed.
I'm working on this - it's something to do with a reactive variable in vue; if i "null" or "[]" the array, it will never populate because at some point vue tracks the variable as "undefined".
And cannot refresh the page (@submit="search" instead of @submit.prevent="search") because it wont ever get data.